requirements.txt
requests
sqlalchemy
mysql-connector-python
python-dotenv
